ingestion-service/requirements.txt
pyzmq>=25.1.0
asyncpg>=0.29.0
questdb>=4.1.0
python-dotenv>=1.0.0